add example training job (#126)
* added example training job

* load data

* rmtree fix

* epoch edit

* fix lint error

* fix lint error

fix line length

fix lint error

+++ truss/truss_handle.py
@@ -250,7 +250,8 @@ class TrussHandle:
         logs_iterator = get_container_logs(container, follow=True, stream=True)
         for log in logs_iterator:
             print(log[1].decode("utf-8"), end="")
-        rmtree(str(self._spec.data_dir))
+        if self._spec.data_dir.exists():
+            rmtree(str(self._spec.data_dir))
         copy_tree_path(output_dir, self._spec.data_dir)
         rmtree(str(output_dir))
         rmtree(str(variables_dir))
